+++ packages/amplify-provider-awscloudformation/src/admin-login.ts
@@ -1,11 +1,10 @@
-import ora from 'ora';
 import util from 'util';
 import readline from 'readline';
 import { Writable } from 'stream';
 import {
   $TSContext, AmplifyError, AMPLIFY_DOCS_URL, open,
 } from 'amplify-cli-core';
-import { printer } from 'amplify-prompts';
+import { printer, AmplifySpinner } from 'amplify-prompts';
 import { adminVerifyUrl, adminBackendMap, isAmplifyAdminApp } from './utils/admin-helpers';  // eslint-disable-line
 import { AdminLoginServer } from './utils/admin-login-server';
 import { AdminAuthPayload } from './utils/auth-types';
@@ -27,11 +26,16 @@ export const adminLoginFlow = async (context: $TSContext, appId: string, envName
   }
 
   const url = adminVerifyUrl(appId, envName, region);
-  printer.info(`Opening link: ${url}`);
-  await open(url, { wait: false }).catch(e => {
-    printer.error(`Failed to open web browser: ${e?.message || e}`);
-  });
-  const spinner = ora('Confirm login in the browser or manually paste in your CLI login key:\n').start();
+  const spinner = new AmplifySpinner();
+
+  try {
+    await open(url, { wait: false });
+    printer.info(`Opening link: ${url}`);
+    spinner.start('Confirm login in the browser or manually paste in your CLI login key:\n');
+  } catch(_) {
+    printer.info(`Could not open ${url} in the current environment`)
+    spinner.start('Manually enter your CLI login key:\n');
+  }
 
   try {
     // spawn express server locally to get credentials
@@ -122,7 +126,8 @@ export const adminLoginFlow = async (context: $TSContext, appId: string, envName
         cancelGetTokenViaServer();
         cancelGetTokenViaPrompt();
       });
-    spinner.succeed('Successfully received Amplify Studio tokens.');
+
+    spinner.stop('Successfully received Amplify Studio tokens.');
   } catch (e) {
     spinner.stop();
     printer.error(`Failed to authenticate with Amplify Studio: ${e?.message || e}`);