+++ packages/amplify-frontend-javascript/CHANGELOG.md
@@ -36,6 +36,15 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [3.7.0-cdkv2.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@3.6.3...amplify-frontend-javascript@3.7.0-cdkv2.0) (2022-11-01)
+## [3.7.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@3.7.0...amplify-frontend-javascript@3.7.1) (2022-11-17)
+
+**Note:** Version bump only for package amplify-frontend-javascript
+
+
+
+
+
+# [3.7.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@3.6.3...amplify-frontend-javascript@3.7.0) (2022-10-27)
 
 
 ### Features