+++ packages/amplify-go-function-runtime-provider/CHANGELOG.md
@@ -36,6 +36,15 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## [2.3.4-cdkv2.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@2.3.3...amplify-go-function-runtime-provider@2.3.4-cdkv2.0) (2022-11-01)
+## [2.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@2.3.4...amplify-go-function-runtime-provider@2.3.5) (2022-11-17)
+
+**Note:** Version bump only for package amplify-go-function-runtime-provider
+
+
+
+
+
+## [2.3.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-go-function-runtime-provider@2.3.3...amplify-go-function-runtime-provider@2.3.4) (2022-10-27)
 
 **Note:** Version bump only for package amplify-go-function-runtime-provider