+++ packages/amplify-graphiql-explorer/CHANGELOG.md
@@ -27,6 +27,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [2.5.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-graphiql-explorer@2.5.0...@aws-amplify/amplify-graphiql-explorer@2.5.1) (2022-11-17)
+
+**Note:** Version bump only for package @aws-amplify/amplify-graphiql-explorer
+
+
+
+
+
 # 2.5.0 (2022-08-02)