+++ packages/amplify-graphql-migration-tests/CHANGELOG.md
@@ -57,6 +57,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [2.3.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.3.3...amplify-graphql-migration-tests@2.3.4) (2022-11-17)
+
+**Note:** Version bump only for package amplify-graphql-migration-tests
+
+
+
+
+
 ## [2.3.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.3.2...amplify-graphql-migration-tests@2.3.3) (2022-11-11)
 
 **Note:** Version bump only for package amplify-graphql-migration-tests