+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -45,6 +45,18 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## [2.5.4-cdkv2.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.5.3...amplify-nodejs-function-template-provider@2.5.4-cdkv2.0) (2022-11-01)
+# [2.6.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.5.4...amplify-nodejs-function-template-provider@2.6.0) (2022-11-17)
+
+
+### Features
+
+* Nodejs graphql IAM template ([#10997](https://github.com/aws-amplify/amplify-cli/issues/10997)) ([880f7eb](https://github.com/aws-amplify/amplify-cli/commit/880f7eb3996133b31c0f498136a66488d1c8ceeb))
+
+
+
+
+
+## [2.5.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.5.3...amplify-nodejs-function-template-provider@2.5.4) (2022-10-27)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider